# Conflicts:
#	client/src/main/java/client/Main.java
#	client/src/main/java/client/scenes/BoardCtrl.java
#	client/src/main/java/client/scenes/CardCtrl.java
#	client/src/main/java/client/scenes/ListOfCardsCtrl.java
#	client/src/main/java/client/scenes/MainScreenCtrl.java
#	client/src/main/java/client/scenes/TagCircleCtrl.java
#	client/src/main/java/client/scenes/TagCtrl.java
#	client/src/main/java/client/utils/ServerUtils.java
#	client/src/main/resources/client/scenes/Board.fxml
#	client/src/main/resources/client/scenes/Card.fxml
#	client/src/main/resources/client/scenes/ConnectToServer.fxml
#	client/src/main/resources/client/scenes/Tag.fxml
#	client/src/main/resources/styles.css
#	quizzzz.trace.db
#	server/src/main/java/server/services/CardService.java

+++ server/src/main/java/server/services/TagService.java
@@ -86,37 +86,24 @@ public class TagService {
      * @param id
      */
     public void deleteTagById(Long id) {
-        tagRepository.getById(id).removeTagsFromCards();
-        tagRepository.deleteById(id);
+        Tag tag = tagRepository.getById(id);
+        if(tag != null) {
+            tag.removeTagsFromCards();
+            tagRepository.deleteById(id);
+        }
     }
 
     /**
      * Edit the name of a tag and store the edited tag in the database
      * @param id
-     * @param newName
+     * @param newTag
      * @return the edited tag
      */
-    public Tag editTagName(Long id, String newName) throws Exception {
-        if(newName == null || newName.isEmpty()) {
-            throw new Exception("Name should not be null or empty.");
-        }
+    public Tag editTag(Long id, Tag newTag) throws Exception {
         Tag tag = getTagById(id);
-        tag.name = newName;
-        return tagRepository.save(tag);
-    }
-
-    /**
-     * Edit the colour of a tag and store the edited tag in the database
-     * @param id
-     * @param newColour
-     * @return the edited tag
-     */
-    public Tag editTagColour(Long id, String newColour) throws Exception {
-        if(newColour == null || newColour.isEmpty()) {
-            throw new Exception("Colour should not be null or empty.");
-        }
-        Tag tag = getTagById(id);
-        tag.colour = newColour;
+        tag.name = newTag.name;
+        tag.colour = newTag.colour;
+        tag.font = newTag.font;
         return tagRepository.save(tag);
     }
